+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix reading dogu config for key updateSiteUrl; #180
+  - entries are now applied in Jenkins
+- Fix multiple JDK 17 version in Jenkins tools; #182
+
+### Changed
+- Update OpenJDK11 version in Jenkins tools
+- Remove plugin-blocklist.json as it can be configured directly via dogu config
 
 ## [v2.462.3-1] - 2024-10-23
 ### Changed