+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v2.462.3-2] - 2024-11-15
 ### Fixed
 - Fix reading dogu config for key updateSiteUrl; #180
   - entries are now applied in Jenkins